Update the internal type representation to support MPTCs https://git.ps.informatik.uni-kiel.de/theses/2021/2021-lkrueger-ma/-/issues/2
Alongside updates necessary to retain the functionality of some functions with multi-parameter type classes, this commit adds a new field to the `Pred` data type. This field should be used to mark implicit class predicates and is needed to retain their property of always being the minimum predicate of a set, as nullary type classes could otherwise cause a violation of this property.