Fix the non-terminating computation occuring in `classDictConstrPredType` in the dictionary translation https://git.ps.informatik.uni-kiel.de/theses/2021/2021-lkrueger-ma/-/issues/6
The variable types used to build the `PredType` of the dictionary data type constructor were constructed with an infinite list of type variable numbers.
While this infinite list did cause no problems when building the super class predicate set, it resulted in a non-terminating execution when building the result type of the constructor.
This error is fixed with this commit by using a limited list of type variable numbers instead. To get the number of necessary type variables, `ModuleIdent` and `TCEnv` had to be added as arguments to this function. Another option would be to use `classArity` which works with the class environment, but this function might be removed since it not strictly necessary.