Fix the `noPolyPred` predicate filter in the instance check https://git.ps.informatik.uni-kiel.de/theses/2021/2021-lkrueger-ma/-/issues/6
When adapting this filter to multi-parameter type classes, it has erroneously been changed from filtering predicates with a type variable applied to other types as their predicate type to filtering predicates which only contain unapplied type variables as predicate types.
This error is fixed with this commit by inverting the filter, so that predicates with at least one applied type variable in their predicate type are filtered.