Fix the unwanted passing of `ICC`-Flags in the type check https://git.ps.informatik.uni-kiel.de/theses/2021/2021-lkrueger-ma/-/issues/6
`ICC`-Flags have been added to mark implicit class constraints, but it has not yet been considered what should happen to these flags when checking expressions containing functions with an implicit class constraint.
Previously, those flags have been passed on, which caused the inferred types of non-class-functions containing class methods to have one or more constraints marked as implicit class constraints, which is not intended.
This error is fixed with this commit by adding a `Bool` parameter to `funType`, the function used to retrieve the types of such functions, including class methods. If set to `True`, the newly added `removeICCFlag` function is applied to the regularly returned type to transform a potential implicit class constraint to a regular predicate. A comment explaing this behaviour has been added.