Fix the unwanted passing of `ICC`-Flags in the dictionary translation https://git.ps.informatik.uni-kiel.de/theses/2021/2021-lkrueger-ma/-/issues/6
The error fixed for the type check in the previous commit did also occur in the dictionary translation. However, in addition to `funType`, it also affected `varType`, which, among other things, is used to retrieve the types of default method implementations which currently have implicit class constraints.
These errors are fixed with this commit by changing `funType` like in the type check and `varType` by adding a wrapper function calling `removeICCFlag` (the preservation of the implicit class constraint flag is not used with this function, therefore an additional parameter is not needed). Comments explaing the behaviour of these functions have been added.