Perform case mode check last
This fixes an issue where the check would wrongly warn (in curry/haskell
mode) on a function declaration like

  neg :: Bool -> Bool
  neg False = True
  neg True  = False

because the `Bool` in the function signature is initially parsed as a
type variable. Since this seems to be disambiguated first in the
TypeSyntaxCheck, we have to make sure that the CaseModeCheck happens
after that. The WarnCheck is performed last, therefore this was not an
issue before.